fix sequence id generator

+++ src/main/java/com/example/sweater/repos/ProductRepo.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ProductRepo extends CrudRepository<Product, Integer> {
+public interface ProductRepo extends CrudRepository<Product, Long> {
 
     List<Product> findAllByOrderByIdDesc();
